Filtering and analysis tools
New
- Added a function (xcorr_microstructure) to cross-correlate the model
estimates of SSA and field measurements. This helps to minimize location
errors and create equivalent positions between the two datasets. Will
write up a demo of this later
- Added filtering in the SMP class (filter_arr) to deal with negative
force values that are showing up with some of the SnowEx SMP units.
Gives two options for now, remove and replace with 0s or remove and
interpolate. All filtering is applied to the subset data and the raw
data is left alone.

Fixes
- The z estimates were missing from the shotNoise variable, added this.
-A few small code changes to reduce memory footprint